fix: ToC challenger 2 code
